+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/ConcordAuthenticatingFilter.java
@@ -35,7 +35,8 @@ public class ConcordAuthenticatingFilter extends AuthenticatingFilter {
      */
     private static final String[] FORCE_BASIC_AUTH_URLS = {
             "/forms/.*",
-            "/api/service/process_portal/.*"
+            "/api/service/process_portal/.*",
+            "/jolokia/.*"
     };
 
     private final ApiKeyDao apiKeyDao;