+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/SecurityModule.java
@@ -21,5 +21,6 @@ public class SecurityModule extends ShiroWebModule {
 
         addFilterChain("/api/**", Key.get(ConcordAuthenticatingFilter.class));
         addFilterChain("/forms/**", Key.get(ConcordAuthenticatingFilter.class));
+        addFilterChain("/jolokia/**", Key.get(ConcordAuthenticatingFilter.class));
     }
 }